+++ RazorChart/RazorChart/Utility/Point.cs
@@ -8,7 +8,7 @@ namespace RazorChart.Utility
 {
     internal class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
     }
 }
